refactor: move execution context related functionality to new file
